Traductor de voz/requirements.txt

streamlit
openai
whisper
TTS
pydub
python-dotenv
torch
